backend/requirements.txt
Flask==2.1.3
PyYAML==6.0.1
Werkzeug==2.0.3
flask-cors
simple-salesforce
pandas
boto3
requests
google-auth
google-auth-oauthlib
google-api-python-client
facebook-business
boto3                # If writing to S3
psycopg2-binary      # If writing to Postgres
sqlalchemy           # If using SQLAlchemy for DB
simplejson